refactor MetricResult to a different file
